Prepare for automated testing
Some changes to the testing logic.

The tester now reacts to the following commands over serial:

'a': disable test auto-start when device is plugged in
't': start test manually
'r': before starting test, drive a "RUN" pin low for 3 seconds (to
     reset the device under test)

automat.py contains a possible way of automatically testing GP2040-CE
devices, it takes an .elf file and a list of configuration binaries
that will each be written to flash and tested. Requires a picoprobe
and openocd.